+++ src/main/kotlin/com/github/pushpavel/autocp/common/res/AutoCpKeys.kt
@@ -5,11 +5,15 @@ import java.util.*
 object AutoCpKeys {
     const val pluginId = "com.github.pushpavel.autocp"
     const val toolWindowSelectedTabIndexKey = "$pluginId.toolWindowSelectedTabIndex"
+    const val autoCpFileVersionNumber = 1
 
     // Command User Macros
     const val inputPathMacro = "@in"
     const val dirPathMacro = "@dir"
     const val dirUnquotedPathMacro = "\$dir"
+
+    // Other Macros
+    const val groupNameMacro = "\$groupName"
 
     // File template keys
     const val fileTemplateName = "CP_TEMPLATE"