+++ src/main/kotlin/com/github/pushpavel/autocp/gather/models/JsonSchema.kt
@@ -1,6 +1,7 @@
 package com.github.pushpavel.autocp.gather.models
 
 import com.github.pushpavel.autocp.database.models.Problem
+import com.github.pushpavel.autocp.database.models.Testcase
 import kotlinx.serialization.Serializable
 
 /**
@@ -38,3 +39,14 @@ data class BatchJson(
     val id: String,
     val size: Int
 )
+
+/**
+ * Json Schema of a Testcase in ProblemJson
+ */
+@Serializable
+data class TestJson(
+    val input: String,
+    val output: String,
+) {
+    fun toTestcase(name: String) = Testcase(name, input, output)
+}