+++ src/main/kotlin/com/github/pushpavel/autocp/settings/generalSettings/AutoCpGeneralSettings.kt
@@ -1,5 +1,6 @@
 package com.github.pushpavel.autocp.settings.generalSettings
 
+import com.github.pushpavel.autocp.common.res.R
 import com.intellij.openapi.components.*
 
 /**
@@ -12,11 +13,23 @@ import com.intellij.openapi.components.*
 @Service
 class AutoCpGeneralSettings : PersistentStateComponent<AutoCpGeneralSettings> {
     var openFilesOnGather = OpenFileOnGather.ONLY_FIRST
+    var fileGenerationRoot = "\$groupName"
+
+
+    fun constructFileGenerationRoot(groupName: String): String {
+        return fileGenerationRoot
+            .replace(
+                R.keys.groupNameMacro,
+                groupName
+                    .replace("[^0-9a-zA-Z_\\- ]".toRegex(), "")
+            )
+    }
 
     override fun getState() = this
 
     override fun loadState(state: AutoCpGeneralSettings) {
         openFilesOnGather = state.openFilesOnGather
+        fileGenerationRoot = state.fileGenerationRoot
     }
 
     companion object {