+++ CHANGELOG.md
@@ -3,6 +3,18 @@
 
 ## [Unreleased]
 
+## v0.7.0-eap.2
+### Added
+- Confirmation dialog before generating files
+- __Java__
+    - New module is created for each solution file generated
+
+### Removed
+- problem gathering success notifications
+
+### Changed
+- Can only enable files with file extensions configured in Settings/Preferences > Tools > AutoCp > Languages
+
 ## v0.7.0-eap.1
 ### Changed
 - Simplified Language settings